+++ packages/api-derive/src/elections/index.ts
@@ -6,3 +6,4 @@ export * from './approvalsOf';
 export * from './approvalsOfAt';
 export * from './info';
 export * from './voters';
+export * from './voterSets';
