+++ packages/api-derive/src/elections/voters.ts
@@ -4,42 +4,34 @@
 
 import { ApiInterfaceRx } from '@polkadot/api/types';
 import { Observable } from 'rxjs';
-import { AccountId, Vector, Option, SetIndex } from '@polkadot/types';
-import { switchMap, map } from 'rxjs/operators';
+import { AccountId, Vector, SetIndex } from '@polkadot/types';
+import { map } from 'rxjs/operators';
 import { drr } from '../util/drr';
+import { voterSets } from './voterSets';
 
 /**
  * @name voters
- * @returns An mapping of all current voter accounts to their voter set index.
+ * @returns An array of all current voters from all sets.
  * @example
  * <BR>
  *
  * ```javascript
  * api.derive.elections.voters((voters) => {
- *   console.log(`ALICE is a voter in the voter set with index ${voters[ALICE].toString()}.`);
+ *   console.log(`There are ${voters.length} current voters.`);
  * });
  * ```
  */
-export function voters (api: ApiInterfaceRx): () => Observable<Record<string, SetIndex>> {
-  return (): Observable<Record<string, SetIndex>> =>
-    api.query.elections.nextVoterSet<SetIndex>().pipe(
-      switchMap((nextVoterSet: SetIndex): Observable<Vector<Option<AccountId>>[]> =>
-        api.query.elections.voters.multi([...Array(+nextVoterSet + 1).keys()].map((_, i): [number] => [i])) as any as Observable<Vector<Option<AccountId>>[]>
-      ),
-      map((voters: Vector<Option<AccountId>>[]): Record<string, SetIndex> =>
-        voters.reduce((result: Record<string, SetIndex>, vec, setIndex): Record<string, SetIndex> => {
-          vec.forEach((e): void => {
-            // re-create the index based on position 0 is [0][0] and likewise
-            // 64 (0..63 in first) is [1][0] (the first index value in set 2)
-            const accountId: AccountId | null = e.unwrapOr(null);
-
-            if (accountId) {
-              result[accountId.toString()] = new SetIndex(setIndex);
-            }
-          });
-
-          return result;
-        }, {})
+export function voters (api: ApiInterfaceRx): () => Observable<Vector<AccountId>> {
+  return (): Observable<Vector<AccountId>> =>
+    voterSets(api)().pipe(
+      map(
+        (voterSets: Record<string, SetIndex>): Vector<AccountId> =>
+          new Vector(
+            AccountId,
+            Object.keys(voterSets).map(
+              address => new AccountId(address)
+            )
+          )
       ),
       drr()
     );
